hybrid_simultanee performs a simultaneous seg - clustering for bivariate signals.

It is an algorithm which combines dynamic programming and the EM algorithm to calculate the MLE of phi and T, which are the mixture parameters and the change point instants. this algorithm is run for a given number of clusters, and estimates the parameters for a segmentation/clustering model with P clusters and 1:Kmax segments

hybrid_simultanee(
  x,
  P,
  Kmax,
  lmin = 3,
  sameSigma = TRUE,
  sameVar.init = FALSE,
  eps = 1e-06,
  lissage = TRUE,
  pureR = FALSE
)

Arguments

x
the two-dimensional signal, one line per dimension

P
the number of classes

Kmax
the maximal number of segments

lmin
minimum length of segment

sameSigma
should segment have the same variance

sameVar.init
sameVar.init

eps
eps

lissage
should likelihood be smoothed

pureR
should algorithm run in full R or use Rcpp speed improvements

Value

a list with Linc, the incomplete loglikelihood =Linc,param=paramtau posterior probability