Internal segmentation/clustering function

Internal segmentation/clustering function

segclust_internal(
  x,
  seg.var = NULL,
  diag.var = NULL,
  order.var = NULL,
  scale.variable = NULL,
  Kmax,
  ncluster = NULL,
  lmin = NULL,
  dat = NULL,
  type = NULL,
  sameSigma = FALSE,
  subsample_over = 1000,
  subsample_by = NA,
  subsample = TRUE,
  ...
)

Arguments

x
data used for segmentation. Supported: data.frame, Move object, ltraj object

seg.var
for behavioral segmentation: names of the variables used for segmentation (either one or two names).

diag.var
for behavioral segmentation: names of the variables on which statistics are calculated.

order.var
for behavioral segmentation: names of the variable with which states are ordered.

scale.variable
should variables be standardized ? (reduced and centered)

Kmax
maximum number of segments.

ncluster
number of cluster into which segments should be grouped. Can be a vector if one want to test several number of clusters.

lmin
minimum length of segments.

dat
bivariate data (one signal per row)

type
type of segmentation. Either "home-range" or "behavior". Changes default values of arguments order, scale.variable in the different functions used on the output. Default for segmentation: "home-range"; default for segmentation/clustering : "behavior".

sameSigma
does segments have same variance ?

subsample_over
over which size should subsampling begin (depending on speed and memory limitations)

subsample_by
override subsample_over to adjust manually subsampling

subsample
if FALSE disable subsample

...
additional arguments given to chooseseg_lavielle