Segmentation of movement data - Generic function

Segmentation of movement data. No clustering. Method available for data.frame, move and ltraj object. The algorithm finds for each number of segment the optimal segmentation using a Dynamic Programming approach. The number of segment is then chosen using Lavielle's (2005) procedure based on locating rupture in the penalized likelihood.

segmentation(x, ...)

# S3 method for data.frame
segmentation(
  x,
  Kmax,
  lmin,
  type = "home-range",
  seg.var,
  diag.var = seg.var,
  order.var = seg.var[1],
  coord.names = c("x", "y"),
  ...
)

# S3 method for Move
segmentation(
  x,
  Kmax,
  lmin,
  type = "home-range",
  seg.var = NULL,
  diag.var = seg.var,
  order.var = seg.var[1],
  coord.names = c("coords.x1", "coords.x2"),
  ...
)

# S3 method for ltraj
segmentation(
  x,
  Kmax,
  lmin,
  type = "home-range",
  seg.var = NULL,
  diag.var = seg.var,
  order.var = seg.var[1],
  coord.names = c("x", "y"),
  ...
)

Arguments

x
data used for segmentation. Supported: data.frame, Move object, ltraj object

...
additional parameters given to segmentation_internal

Kmax
maximum number of segments.

lmin
minimum length of segments.

type
type of segmentation. Either "home-range" or "behavior". Changes default values of arguments order, scale.variable in the different functions used on the output. Default for segmentation: "home-range"; default for segmentation/clustering : "behavior".

seg.var
for behavioral segmentation: names of the variables used for segmentation (either one or two names).

diag.var
for behavioral segmentation: names of the variables on which statistics are calculated.

order.var
for behavioral segmentation: names of the variable with which states are ordered.

coord.names
for home-range segmentation and data.frame, names of coordinates. Default x and y. Not needed for move and ltraj objects

Value

a segmentation-class object

Examples